InCryptoTransport::partial_read: result type is now size_t

+++ src/transport/in_crypto_transport.hpp
@@ -163,12 +163,12 @@ public:
         this->open(pathname);
     }
 
-    int partial_read(uint8_t * buffer, size_t len) __attribute__ ((warn_unused_result))
+    size_t partial_read(uint8_t * buffer, size_t len) __attribute__ ((warn_unused_result))
     {
         return this->do_partial_read(buffer, len);
     }
 
-    int partial_read(char * buffer, size_t len) __attribute__ ((warn_unused_result))
+    size_t partial_read(char * buffer, size_t len) __attribute__ ((warn_unused_result))
     {
         return this->do_partial_read(reinterpret_cast<uint8_t*>(buffer), len);
     }
@@ -367,7 +367,7 @@ private:
         }
     }
 
-    int do_partial_read(uint8_t * buffer, size_t len)
+    size_t do_partial_read(uint8_t * buffer, size_t len)
     {
         if (this->eof){
             return 0;
@@ -461,7 +461,7 @@ private:
                 }
                 return len;
             }
-            unsigned int remaining_len = len;
+            size_t remaining_len = len;
             if (this->raw_size - this->clear_pos > 0){
                 ::memcpy(&buffer[0], &this->clear_data[this->clear_pos], this->raw_size - this->clear_pos);
                 remaining_len -= this->raw_size - this->clear_pos;
@@ -495,11 +495,11 @@ private:
 
     Read do_atomic_read(uint8_t * buffer, size_t len) override
     {
-        int res = do_partial_read(buffer, len);
-        if ((res != 0) && (res != int(len))) {
+        size_t res = do_partial_read(buffer, len);
+        if (res != 0 && res != len) {
             throw Error(ERR_TRANSPORT_READ_FAILED, 0);
         }
-        return res == int(len) ? Read::Ok : Read::Eof;
+        return res == len ? Read::Ok : Read::Eof;
     }
 
     std::size_t get_file_len(char const * pathname)
